+++ services/catalog/tests/unit/test_services_director.py
@@ -30,6 +30,7 @@ def app_environment(
 
 
 async def test_director_client_setup(
+    with_disabled_rabbitmq_and_rpc: None,
     mocked_director_service_api: MockRouter,
     app: FastAPI,
 ):
